Rename `persistable` to `insertable`
Because `persistable` is the wrong name for a module that only cares
about inserts, and I always look for the file in the wrong place.

I think it's also about time for us to start using deprecation warnings
when we can, and this is one place that we can easily do it.
