+++ bin/test
@@ -8,7 +8,7 @@ elif [ "$1" == "integration" ]; then
 elif [ "$1" == "compile" ]; then
   (cd diesel_compile_tests && cargo test)
 else
-  (cd diesel && cargo test --features "unstable chrono sqlite")
+  (cd diesel && cargo test --no-default-features --features "unstable chrono sqlite postgres")
   (cd diesel_cli && cargo test --features "sqlite" --no-default-features)
   (cd diesel_tests && DATABASE_URL=/tmp/test.db cargo test --features "unstable_sqlite" --no-default-features)
   (cd examples && ./test_nightly)
